Add snapshots and smart constructor for Address
