+++ libs/domain/feature-funds/src/FundSnapshot.ts
@@ -1,0 +1,6 @@
+/**
+ * Base type for a temporal snapshot of a fund.
+ */
+export type FundSnapshot = {
+  timestamp: Date;
+};
