+++ libs/domain/feature-funds/src/Token.ts
@@ -4,6 +4,12 @@ import { Address } from './Address';
  * Represents a token on the Ethereum blockchain.
  */
 export type Token = {
+  /**
+   * The address where this token is deployed. This is important
+   * because `name` and `symbol` are not unique according to the
+   * ERC-20 specification.
+   */
+  address: Address;
   name: string;
   symbol: string;
   decimals: number;
@@ -12,12 +18,6 @@ export type Token = {
    * to create tagged unions.
    */
   kind: string;
-  /**
-   * The address where this token is deployed. This is important
-   * because `name` and `symbol` are not unique according to the
-   * ERC-20 specification.
-   */
-  address: Address;
 };
 
 /**
